fix stack types

+++ include/stack.h
@@ -6,12 +6,12 @@
 /*
  * Push a value onto the Stack of the VM
  */
-int push_stack(struct kvm_vm *, struct kvm_vcpu *, uint16_t);
+int push_stack(struct kvm_vm *, struct kvm_vcpu *, uint64_t);
 
 /*
  * Pop a value from the VM's Stack
  */
-uint16_t pop_stack(struct kvm_vm *, struct kvm_vcpu *);
+uint64_t pop_stack(struct kvm_vm *, struct kvm_vcpu *);
 
 /*
  * \brief Expand the Stack by one Frame
